Modify default namespace settings for Prometheus
Before this commit, Prometheus had the following three settings for each CR:
```
<serviceMonitor|podMonitor|rule>NamespaceSelector: {}
<serviceMonitor|podMonitor|rule>SelectorNilUsesHelmValues: true
<serviceMonitor|podMonitor|rule>Selector: {}
```
The first setting, when nil, indicates that we only select resources created in the deployment namespace.
The second setting, when true, indicates that we only select resources with the Helm release label.
The third setting, when nil, indicates that we should use the second setting.

This behavior has now been modified to the following:
```
<serviceMonitor|podMonitor|rule>NamespaceSelector:
    any: true
<serviceMonitor|podMonitor|rule>SelectorNilUsesHelmValues: false
<serviceMonitor|podMonitor|rule>Selector: {}
```

This indicates that we should select resources from all namespaces with any labels.